refactor: `UpdateNode`
